# schemas/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Backend API",
    "description": "Express server OAS"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "devServer"
    }
  ],
...
  "paths": {
...
    "/api/v1/cv/publications": {
      "get": {
        "tags": [
          "CV"
        ],
        "summary": "returns list of publications with static contents",
        "description": "/api/v1/cv/publications",
        "parameters": [
          {
            "name": "host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/cv/publications/{filename}": {
      "get": {
        "tags": [
          "CV"
        ],
        "summary": "send file with pdf format",
        "description": "/api/v1/cv/publications/:filename",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}